+++ include/aligator/gar/parallel-solver.hxx
@@ -4,21 +4,38 @@
 
 #include "aligator/gar/parallel-solver.hpp"
 #include "aligator/gar/block-tridiagonal.hpp"
-#include "aligator/gar/work.hpp"
 #include "aligator/gar/lqr-problem.hpp"
-#include "aligator/utils/mpc-util.hpp"
+// #include "aligator/utils/mpc-util.hpp"
 #include "aligator/tracy.hpp"
 #include "aligator/utils/exceptions.hpp"
-
 #include "aligator/threads.hpp"
 
+#include <numeric>
+
 namespace aligator::gar {
+struct workrange_t {
+  uint beg;
+  uint end;
+};
+
+/// @brief Get a balanced work range corresponding to a horizon @p horz, thread
+/// ID @p tid, and number of threads @p num_threads.
+constexpr workrange_t get_work(uint horz, uint thread_id, uint num_threads) {
+  uint start = thread_id * (horz + 1) / num_threads;
+  uint stop = (thread_id + 1) * (horz + 1) / num_threads;
+  assert(stop <= horz + 1);
+  return {start, stop};
+}
 
 #ifdef ALIGATOR_MULTITHREADING
 template <typename Scalar>
 ParallelRiccatiSolver<Scalar>::ParallelRiccatiSolver(
     LqrProblemTpl<Scalar> &problem, const uint num_threads)
     : Base()
+    , condensedKktSystem(problem.get_allocator())
+    , condensedKktRhs(problem.get_allocator())
+    , condensedKktSolution(problem.get_allocator())
+    , condensedErr(problem.get_allocator())
     , numThreads(num_threads)
     , problem_(&problem) {
   ALIGATOR_TRACY_ZONE_SCOPED;
@@ -28,36 +45,40 @@ ParallelRiccatiSolver<Scalar>::ParallelRiccatiSolver(
         __FUNCTION__, numThreads);
   }
 
-  uint N = (uint)problem.horizon();
-  for (uint i = 0; i < numThreads; i++) {
-    auto [start, end] = get_work(N, i, numThreads);
-    allocateLeg(start, end, i == (numThreads - 1));
-  }
-
-  std::vector<long> dims{problem.nc0(), problem.stages.front().nx};
-  for (uint i = 0; i < numThreads - 1; i++) {
-    auto [i0, i1] = get_work(N, i, numThreads);
-    dims.push_back(problem.stages[i0].nx);
-    dims.push_back(problem.stages[i1 - 1].nx);
-  }
-  condensedKktRhs = BlkVec(dims);
-  condensedKktSolution = condensedKktRhs;
-  condensedErr = condensedKktRhs;
-  initializeTridiagSystem(dims);
-
-  assert(datas.size() == (N + 1));
+  this->initialize();
 }
 
-template <typename Scalar>
-void ParallelRiccatiSolver<Scalar>::allocateLeg(uint start, uint end,
-                                                bool last_leg) {
-  ALIGATOR_TRACY_ZONE_SCOPED;
-  for (uint t = start; t < end; t++) {
-    KnotType &knot = problem_->stages[t];
-    if (!last_leg)
-      knot.addParameterization(knot.nx);
-    datas.emplace_back(knot.nx, knot.nu, knot.nc, knot.nx2, knot.nth);
+template <typename Scalar> void ParallelRiccatiSolver<Scalar>::initialize() {
+  const auto allocate_leg = [this](uint start, uint end, bool last_leg) {
+    uint nth = problem_->stages[end - 1].nx2;
+    for (uint t = start; t < end; t++) {
+      KnotType &knot = problem_->stages[t];
+      if (!last_leg)
+        knot.addParameterization(nth);
+      datas.emplace_back(knot.nx, knot.nu, knot.nc, knot.nx2, knot.nth);
+    }
+  };
+
+  const uint N = (uint)problem_->horizon();
+  for (uint i = 0; i < numThreads; i++) {
+    auto [i0, i1] = get_work(N, i, numThreads);
+    allocate_leg(i0, i1, i == (numThreads - 1));
   }
+
+  rhsDims_ = {problem_->nc0(), problem_->stages[0].nx};
+  for (uint i = 0; i < numThreads - 1; i++) {
+    auto [i0, i1] = get_work(N, i, numThreads);
+    rhsDims_.push_back(problem_->stages[i0].nx);
+    rhsDims_.push_back(problem_->stages[i1 - 1].nx);
+  }
+  long condensed_total_dim =
+      std::accumulate(rhsDims_.begin(), rhsDims_.end(), 0l);
+  condensedKktRhs.setZero(condensed_total_dim);
+  condensedKktSolution = condensedKktRhs;
+  condensedErr = condensedKktRhs;
+  initializeTridiagSystem();
+
+  assert(datas.size() == (N + 1));
 }
 
 template <typename Scalar>
@@ -95,26 +116,34 @@ void ParallelRiccatiSolver<Scalar>::assembleCondensedSystem(
     subdiagonal[i] = superdiagonal[i].transpose();
   }
 
-  condensedKktRhs[0] = -problem_->g0;
-  condensedKktRhs[1] = -datas[0].vm.vx;
+  BlkView rhs_view{condensedKktRhs, rhsDims_};
+  rhs_view.blockSegment(0) = -problem_->g0;
+  rhs_view.blockSegment(1) = -datas[0].vm.vx;
 
   for (uint i = 0; i < numThreads - 1; i++) {
     auto [i0, i1] = get_work(N, i, numThreads);
     uint ip1 = i + 1;
-    condensedKktRhs[2 * ip1] = -datas[i0].vm.vt;
-    condensedKktRhs[2 * ip1 + 1] = -datas[i1].vm.vx;
+    rhs_view.blockSegment(2 * ip1) = -datas[i0].vm.vt;
+    rhs_view.blockSegment(2 * ip1 + 1) = -datas[i1].vm.vx;
   }
 }
 
 template <typename Scalar>
 bool ParallelRiccatiSolver<Scalar>::backward(const Scalar mueq) {
-
   ALIGATOR_NOMALLOC_SCOPED;
   ALIGATOR_TRACY_ZONE_SCOPED_N("parallel_backward");
-  auto N = static_cast<uint>(problem_->horizon());
+
+  const auto configure_knot = [](KnotType &knot) {
+    knot.Gx = knot.A.transpose();
+    knot.Gu = knot.B.transpose();
+    knot.Gth.setZero();
+    knot.gamma = knot.f;
+  };
+
+  const uint N = static_cast<uint>(problem_->horizon());
   for (uint i = 0; i < numThreads - 1; i++) {
     uint end = get_work(N, i, numThreads).end;
-    setupKnot(problem_->stages[end - 1]);
+    configure_knot(problem_->stages[end - 1]);
   }
   Eigen::setNbThreads(1);
   aligator::omp::set_default_options(numThreads, false);
@@ -143,33 +172,33 @@ bool ParallelRiccatiSolver<Scalar>::backward(const Scalar mueq) {
     condensedKktSystem.upFacs = condensedKktSystem.subdiagonal;
 
     // This routine may have accuracy issues
-    symmetricBlockTridiagSolve<ArMat>(
-        condensedKktSystem.subdiagonal, condensedKktSystem.diagonal,
-        condensedKktSystem.superdiagonal, condensedKktSolution,
-        condensedKktSystem.ldlt);
+    BlkView kkt_sol_view{condensedKktSolution, rhsDims_};
+    symmetricBlockTridiagSolve<ArMat>(condensedKktSystem.subdiagonal,
+                                      condensedKktSystem.diagonal,
+                                      condensedKktSystem.superdiagonal,
+                                      kkt_sol_view, condensedKktSystem.ldlt);
     condensedKktSystem.diagonalFacs.swap(condensedKktSystem.diagonal);
     condensedKktSystem.upFacs.swap(condensedKktSystem.subdiagonal);
 
     // iterative refinement
-    constexpr int maxRefinementSteps = 5;
-    for (int i = 0; i < maxRefinementSteps; i++) {
+    BlkView err_rhs_view{condensedErr, rhsDims_}; // blocked view of the rhs
+    for (uint i = 0; i < maxRefinementSteps; i++) {
       // 1. compute residual into rhs
-      blockTridiagMatMul<ArMat>(condensedKktSystem.subdiagonal,
-                                condensedKktSystem.diagonal,
-                                condensedKktSystem.superdiagonal,
-                                condensedKktSolution, condensedErr, -1.0);
-      condensedErr.matrix() *= -1;
+      blockTridiagMatMul<ArMat>(
+          condensedKktSystem.subdiagonal, condensedKktSystem.diagonal,
+          condensedKktSystem.superdiagonal, kkt_sol_view, err_rhs_view, -1.0);
+      condensedErr *= -1;
 
-      Scalar resdl = math::infty_norm(condensedErr.matrix());
+      Scalar resdl = math::infty_norm(condensedErr);
       if (resdl <= condensedThreshold)
         return true;
 
       // 2. perform refinement step and swap
       blockTridiagRefinementStep<ArMat>(condensedKktSystem.upFacs,
                                         condensedKktSystem.superdiagonal,
-                                        condensedKktSystem.ldlt, condensedErr);
-      condensedKktSolution.matrix() += condensedErr.matrix();
-      condensedErr = condensedKktRhs.matrix();
+                                        condensedKktSystem.ldlt, err_rhs_view);
+      condensedKktSolution += condensedErr;
+      condensedErr = condensedKktRhs;
     }
   }
 
@@ -183,10 +212,11 @@ bool ParallelRiccatiSolver<Scalar>::forward(
   ALIGATOR_NOMALLOC_SCOPED;
   ALIGATOR_TRACY_ZONE_SCOPED_N("parallel_forward");
   uint N = (uint)problem_->horizon();
+  BlkView sol_view{condensedKktSolution, rhsDims_};
   for (uint i = 0; i < numThreads; i++) {
     uint i0 = get_work(N, i, numThreads).beg;
-    lbdas[i0] = condensedKktSolution[2 * i];
-    xs[i0] = condensedKktSolution[2 * i + 1];
+    lbdas[i0] = sol_view.blockSegment(2 * i);
+    xs[i0] = sol_view.blockSegment(2 * i + 1);
   }
   Eigen::setNbThreads(1);
   const auto &stages = problem_->stages;
@@ -213,31 +243,23 @@ bool ParallelRiccatiSolver<Scalar>::forward(
 }
 
 template <typename Scalar>
-void ParallelRiccatiSolver<Scalar>::cycleAppend(const KnotType &knot) {
-  rotate_vec_left(datas, 0, 1);
-  datas[problem_->horizon() - 1ul] =
-      StageFactor<Scalar>(knot.nx, knot.nu, knot.nc, knot.nx2, knot.nth);
-  rotate_vec_left(condensedKktSystem.subdiagonal);
-  rotate_vec_left(condensedKktSystem.diagonal);
-  rotate_vec_left(condensedKktSystem.superdiagonal);
-  rotate_vec_left(condensedKktSystem.diagonalFacs);
-  rotate_vec_left(condensedKktSystem.upFacs);
-  rotate_vec_left(condensedKktSystem.ldlt);
-
-  auto [i0, i1] = get_work(problem_->horizon(), numThreads - 2u, numThreads);
-  uint dim0 = problem_->stages[i0].nx;
-  uint dim1 = problem_->stages[i1 - 1u].nx;
-  condensedKktSystem.subdiagonal.back().setZero(dim1, dim0);
-  condensedKktSystem.diagonal.back().setZero(dim1, dim1);
-  condensedKktSystem.diagonalFacs.back().setZero(dim1, dim1);
-  condensedKktSystem.upFacs.back().setZero(dim1, dim1);
-  condensedKktSystem.ldlt.back() = BunchKaufman<MatrixXs>(dim1);
+void ParallelRiccatiSolver<Scalar>::cycleAppend(const KnotType &) {
+  datas.clear();
+  condensedKktSystem.subdiagonal.clear();
+  condensedKktSystem.diagonal.clear();
+  condensedKktSystem.superdiagonal.clear();
+  condensedKktSystem.diagonalFacs.clear();
+  condensedKktSystem.upFacs.clear();
+  condensedKktSystem.ldlt.clear();
+  problem_->addParameterization(0);
+  // just... reinitialise everything
+  this->initialize();
 };
 
 template <typename Scalar>
-void ParallelRiccatiSolver<Scalar>::initializeTridiagSystem(
-    const std::vector<long> &dims) {
+void ParallelRiccatiSolver<Scalar>::initializeTridiagSystem() {
   ALIGATOR_TRACY_ZONE_SCOPED;
+  const auto &dims = rhsDims_;
 
   condensedKktSystem.subdiagonal.reserve(dims.size() - 1);
   condensedKktSystem.diagonal.reserve(dims.size());
@@ -246,7 +268,7 @@ void ParallelRiccatiSolver<Scalar>::initializeTridiagSystem(
   condensedKktSystem.upFacs.reserve(dims.size());
   condensedKktSystem.ldlt.reserve(dims.size());
 
-  const auto emplace_factor = [](condensed_system_t &f, Eigen::Index dim) {
+  const auto emplace_factor = [](CondensedKkt &f, Eigen::Index dim) {
     f.diagonalFacs.emplace_back(dim, dim);
     f.upFacs.emplace_back(dim, dim);
     f.ldlt.emplace_back(dim);